Work in progress changing "Int" and "UInt" types

+++ javatools/src/main/java/org/xvm/asm/constants/ByteConstant.java
@@ -530,7 +530,7 @@ public class ByteConstant
         ConstantPool pool = getConstantPool();
         if (typeOut.equals(pool.typeInt()))
             {
-            return IntConstant.toIntConstant(Format.Int, new PackedInteger(m_nVal), pool);
+            return IntConstant.toIntConstant(Format.Int64, new PackedInteger(m_nVal), pool);
             }
         else if (typeOut.equals(pool.typeInt8()))
             {
@@ -558,7 +558,7 @@ public class ByteConstant
             }
         else if (typeOut.equals(pool.typeUInt()))
             {
-            return IntConstant.toIntConstant(Format.UInt, new PackedInteger(m_nVal), pool);
+            return IntConstant.toIntConstant(Format.UInt64, new PackedInteger(m_nVal), pool);
             }
         else if (typeOut.equals(pool.typeBit()))
             {
