+++ javatools/src/main/java/org/xvm/compiler/ast/ToIntExpression.java
@@ -140,7 +140,7 @@ public class ToIntExpression
         String       sFormat = expr.getType().getEcstasyClassName(); // REVIEW CP
         return switch (sFormat)
             {
-            case "numbers.Int"     -> pool.ensureIntConstant(pint, Format.Int);
+            case "numbers.Int"     -> pool.ensureIntConstant(pint, Format.Int64);
             case "numbers.Int8"    -> pool.ensureByteConstant(Format.Int8, pint.getInt());
             case "numbers.UInt8"   -> pool.ensureByteConstant(Format.UInt8, pint.getInt());
             case "numbers.Int16"   -> pool.ensureIntConstant(pint, Format.Int16);
@@ -148,7 +148,7 @@ public class ToIntExpression
             case "numbers.Int64"   -> pool.ensureIntConstant(pint, Format.Int64);
             case "numbers.Int128"  -> pool.ensureIntConstant(pint, Format.Int128);
             case "numbers.IntN"    -> pool.ensureIntConstant(pint, Format.IntN);
-            case "numbers.UInt"    -> pool.ensureIntConstant(pint, Format.UInt);
+            case "numbers.UInt"    -> pool.ensureIntConstant(pint, Format.UInt64);
             case "numbers.UInt16"  -> pool.ensureIntConstant(pint, Format.UInt16);
             case "numbers.UInt32"  -> pool.ensureIntConstant(pint, Format.UInt32);
             case "numbers.UInt64"  -> pool.ensureIntConstant(pint, Format.UInt64);
